requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
Django==3.1.3
django-crispy-forms==1.9.2
django-phone-field==1.8.1
filelock==3.0.12
isort==4.3.21
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.3.2
mccabe==0.6.1
networkx==2.5
numpy==1.19.2
Pillow==7.2.0
pycodestyle==2.5.0
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
virtualenv==20.0.31
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
wrapt==1.12.1
